+++ src/main/java/cn/nukkit/block/StoneBricks.java
@@ -9,10 +9,10 @@ import cn.nukkit.math.AxisAlignedBB;
  * Nukkit Project
  */
 public class StoneBricks extends Solid {
-    public static final short NORMAL = 0;
-    public static final short MOSSY = 1;
-    public static final short CRACKED = 2;
-    public static final short CHISELED = 3;
+    public static final int NORMAL = 0;
+    public static final int MOSSY = 1;
+    public static final int CRACKED = 2;
+    public static final int CHISELED = 3;
 
     protected int id = STONE_BRICKS;
